Points to drawing order wasn't backtracking
The previous approach to collecting points to drawing order didn't
account correctly for 'dead ends', i.e. pixels without a neighbor like
the antenna in the car sprite included in assets directory. The code had
what turned out to be a fragile workaround that really only worked for
the car sprite: jumbling the order of the initial points a bit.

This commit gets rid of the workaround, and handles backtracking by
walking backward when we are at a pixel with no neighbor but we're not
out of points.